moved the file-dialog module to the filesystem from the workspace extension
