+++ examples/browser/src/frontend/main.ts
@@ -21,7 +21,6 @@ import 'theia-core/src/application/browser/style/index.css';
 import 'theia-core/src/monaco/browser/style/index.css';
 import 'theia-core/src/filesystem/browser/style/index.css';
 import 'theia-core/src/terminal/browser/terminal.css';
-import 'theia-core/src/workspace/browser/file-dialog/file-dialog.css';
 import 'font-awesome/css/font-awesome.min.css';
 
 // terminal extension
