+++ src/filesystem/browser/style/file-dialog.css
