+++ src/filesystem/browser/style/index.css
@@ -1,14 +1,9 @@
-.theia-FileStatIcon {
-    width: 16px;
-    height: 16px;
-    margin-top: 1px;
-    margin-right: 3px;
-}
+/*
+ * Copyright (C) 2017 TypeFox and others.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ */
 
-.theia-FileStatNode:not(.theia-DirNode) > .theia-TreeNodeCaption > .theia-FileStatIcon {
-    background: url('../icons/path.svg') center center no-repeat;
-}
-
-.theia-FileStatNode.theia-DirNode > .theia-TreeNodeCaption > .theia-FileStatIcon {
-    background: url('../icons/dir.svg') center center no-repeat;
-}
+@import './file-tree.css';
+@import './file-dialog.css';
